requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
bidict==0.23.1
black==24.10.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dnspython==2.7.0
eventlet==0.38.0
flake8==7.1.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-sock==0.7.0
Flask-SocketIO==5.4.1
fonttools==4.55.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.2
paramiko==3.5.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
psutil==6.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-ldap==3.4.4
python-socketio==5.11.4
pytz==2024.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
scapy==2.6.1
scipy==1.14.1
simple-websocket==1.1.0
six==1.16.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
